PAY-7517: Resolve problem  for users deleted by IDAM team         [PAY-7517] (https://tools.hmcts.net/jira/browse/PAY-7517).
